Load settings and runtime states from config
In origin, the settings and runtime states are injected into the code
by meta . This method is not stable, hard to maintain and not cross
platforms. Loading these data from config file is a suitable altertive.

In the future, we may move the runtime states to cache to stop the
program modifying the config file at runtime.